Album에 tracks 추가 #3

+++ src/main/kotlin/com/moontomi/server/core/album/AlbumService.kt
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Service
 
 @Service
 class AlbumService {
-    fun findById(id: Long) {
+    fun findById(id: Int) {
     }
 
     fun insert() {
